move deprecated API to separate module
